Add fuzzing target to Makefile
