+++ examples/fuzz_isort/fuzz.py
@@ -1,16 +1,16 @@
 # mypy: disable-error-code="attr-defined"
 
-import zlib
-
 from cobrafuzz.main import CobraFuzz
 
 
 @CobraFuzz
 def fuzz(buf: bytes) -> None:
-    try:  # noqa: SIM105
-        zlib.decompress(buf)
-    except zlib.error:
-        pass
+    import contextlib
+
+    import isort
+
+    with contextlib.suppress(UnicodeDecodeError):
+        isort.code(buf.decode("ascii"))
 
 
 if __name__ == "__main__":
